+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   Rendering issue in attached image preview when sending message on web.
 * **Feat**: [420](https://github.com/SimformSolutionsPvtLtd/chatview/pull/420) Added support for
   `playerMode` in `VoiceMessageConfiguration` with `single` and `multi`.
+* **Feat**: [429](https://github.com/SimformSolutionsPvtLtd/chatview/pull/429) Added optional `headerSlivers`
+  to `ChatList` for better customization.
 
 ## [3.0.0]